requirements.txt
aniso8601==8.0.0
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Django==3.0
djangorestframework==3.10.3
Flask==1.1.1
Flask-RESTful==0.3.7
flask-restful-swagger==0.20.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
Werkzeug==0.16.0
